+++ firebase-api-library/src/main/java/com/fazemeright/library/api/domain/authentication/UserAuthResult.kt
@@ -1,4 +1,4 @@
-package com.fazemeright.library.api
+package com.fazemeright.library.api.domain.authentication
 
 interface UserAuthResult {
     //  TODO: Replace with custom user object
