--- firebase-api-library/src/main/java/com/fazemeright/library/api/firebase/OnCompleteForOnTaskAuthResultAdapter.kt
@@ -1,16 +1,0 @@
-package com.fazemeright.library.api.firebase
-
-/*
-class OnCompleteForOnTaskAuthResultAdapter(
-        private val onTaskCompleteListener: OnTaskCompleteListener<UserAuthResult?>) : OnCompleteListener<AuthResult> {
-    override fun onComplete(task: Task<AuthResult>) {
-        if (task.isSuccessful) {
-            onTaskCompleteListener
-                    .onComplete(withData(UserAuthResultAdapterForAuthResult(task.result)))
-        } else {
-            onTaskCompleteListener
-                    .onComplete(Result.exception(task.exception))
-        }
-    }
-
-}*/
