--- firebase-api-library/src/main/java/com/fazemeright/library/api/firebase/OnTaskCompleteAdapterForOnComplete.kt
@@ -1,13 +1,0 @@
-package com.fazemeright.library.api.firebase
-
-/*
-class OnTaskCompleteAdapterForOnComplete<T>(
-        private val onTaskCompleteListener: OnTaskCompleteListener<T?>) : OnCompleteListener<T> {
-    override fun onComplete(task: Task<T>) {
-        if (task.isSuccessful) {
-            onTaskCompleteListener.onComplete(withData(task.result))
-        } else {
-            onTaskCompleteListener.onComplete(Result.exception(task.exception))
-        }
-    }
-}*/
